requirements.txt
Authlib==1.3.2
Flask==3.1.0
Flask_Cors==5.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
